% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{stackedSVD}
\alias{stackedSVD}
\title{Stacked SVD}
\usage{
stackedSVD(X.list, rank = "auto", max.rank = 50)
}
\arguments{
\item{X.list}{A list of numeric matrices with matching column dimensions.}

\item{rank}{Desired rank or "auto" for automatic selection.}

\item{max.rank}{Maximum allowable rank for truncated SVD.}
}
\value{
A matrix of right singular vectors.
}
\description{
Performs truncated SVD on the row-stacked version of a list of matrices.
}
