% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{stackedSVDWeighted}
\alias{stackedSVDWeighted}
\title{SVD Stacked with Weighting}
\usage{
stackedSVDWeighted(X.list, rank, max.rank = 50)
}
\arguments{
\item{X.list}{A list of numeric matrices with matching column dimensions.}

\item{rank}{Desired number of components to keep.}

\item{max.rank}{Maximum allowable rank for truncated SVD.}
}
\value{
A matrix of consensus right singular vectors.
}
\description{
Computes right singular vectors using a weighted stacking procedure based on theoretical theta values.
}
